+++ src/de.rs
@@ -1,3 +1,5 @@
+//! Helpers for annotating types to deserialize from ITF values.
+
 use serde::de::DeserializeOwned;
 
 use crate::Value;
@@ -6,12 +8,12 @@ mod error;
 pub use error::Error;
 
 mod helpers;
-pub use helpers::{As, Integer, Same};
+pub use helpers::{As, Integer, Option, Result, Same};
 
 mod deserializer;
 
 #[doc(hidden)]
-pub fn decode_value<T>(value: Value) -> Result<T, Error>
+pub fn decode_value<T>(value: Value) -> std::result::Result<T, Error>
 where
     T: DeserializeOwned,
 {