+++ src/state.rs
@@ -1,3 +1,5 @@
+//! Defines the states contained within an ITF trace.
+
 use std::collections::BTreeMap;
 
 use serde::de::DeserializeOwned;